refactor (remove Lottery) : Lottery 객체가 쓸모 없어져서 삭제
