+++ src/main/java/lotto/domain/Lotteries.java
@@ -1,25 +1,35 @@
 package lotto.domain;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 
 public class Lotteries {
-    private final List<Lottery> lotteries;
 
-    public Lotteries(final List<Lottery> lotteries) {
-        this.lotteries = lotteries;
+    private List<LottoNumbers> lottoNumbers;
+
+    public Lotteries(final List<LottoNumbers> lottoNumbers) {
+        this.lottoNumbers = Collections.unmodifiableList(lottoNumbers);
     }
 
     public int count() {
-        return lotteries.size();
+        return lottoNumbers.size();
     }
 
     public List<LottoNumbers> getLottoNumbers() {
-        List<LottoNumbers> lottoNumbers = new ArrayList<>();
-        for (Lottery lottery : lotteries) {
-            lottoNumbers.add(lottery.getLottoNumbers());
-        }
-        return Collections.unmodifiableList(lottoNumbers);
+        return lottoNumbers;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        final Lotteries lotteries = (Lotteries) o;
+        return Objects.equals(lottoNumbers, lotteries.lottoNumbers);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(lottoNumbers);
     }
 }
