+++ src/main/java/lotto/domain/Lotto.java
@@ -5,14 +5,13 @@ import java.util.List;
 
 public class Lotto {
     public Lotteries buy(final Price price) {
-        List<Lottery> lotteries = publishLottery(price);
-        return new Lotteries(lotteries);
+        return new Lotteries(publishLottery(price));
     }
 
-    private List<Lottery> publishLottery(final Price price) {
-        List<Lottery> lotteries = new ArrayList<>();
+    private List<LottoNumbers> publishLottery(final Price price) {
+        List<LottoNumbers> lotteries = new ArrayList<>();
         for (int i = 0; i < price.lotteryCount(); i++) {
-            lotteries.add(new Lottery(LottoNumberGenerator.generate()));
+            lotteries.add(LottoNumberGenerator.generate());
         }
         return lotteries;
     }
